+++ packages/core/src/__tests__/constants.spec.ts
@@ -1,5 +1,6 @@
 // Copyright 2020 Cognite AS
 
+import { expect, test } from 'vitest';
 import { BASE_URL } from '../constants';
 
 test('BASE_URL', () => {