+++ packages/core/src/__tests__/graphUtils.unit.spec.ts
@@ -1,6 +1,7 @@
 // Copyright 2020 Cognite AS
 
-import { Node, topologicalSort } from '../graphUtils';
+import { describe, expect, test } from 'vitest';
+import { topologicalSort } from '../graphUtils';
 
 describe('Graph utils', () => {
   describe('topologicalSort', () => {
@@ -44,7 +45,7 @@ describe('Graph utils', () => {
       expect(() =>
         topologicalSort([nodeA, nodeAA, nodeAAA])
       ).toThrowErrorMatchingInlineSnapshot(
-        `"Impossible to topological sort nodes"`
+        `[Error: Impossible to topological sort nodes]`
       );
     });
   });