+++ packages/stable/src/__tests__/api/instances.int.spec.ts
@@ -1,7 +1,6 @@
 // Copyright 2024 Cognite AS
 
-import { ViewReference } from 'stable/src/types';
-import CogniteClient from '../../cogniteClient';
+import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest';
 import { setupLoggedInClient } from '../testUtils';
 
 type SpaceDefinition = {
@@ -81,7 +80,7 @@ describe('Instances integration test', () => {
     client = setupLoggedInClient();
     await upsertSpace(client, testSpace);
     await upsertDescribables(client, [describable1, describable2], view);
-  });
+  }, 10000);
 
   afterAll(async () => {
     await client.instances.delete([
@@ -96,7 +95,7 @@ describe('Instances integration test', () => {
         space: testSpace.space,
       },
     ]);
-  });
+  }, 10000);
 
   test('list nodes from a single view with limit 2', async () => {
     const response = await client.instances.list({
@@ -219,24 +218,29 @@ describe('Instances integration test', () => {
   });
 
   test('search with filter', async () => {
-    const response = await client.instances.search({
-      view,
-      filter: {
-        prefix: {
-          property: ['title'],
-          value: 'titl',
-        },
+    vi.waitFor(
+      async () => {
+        const response = await client.instances.search({
+          view,
+          filter: {
+            prefix: {
+              property: ['title'],
+              value: 'titl',
+            },
+          },
+          limit: 1,
+        });
+        expect(response.items).toHaveLength(1);
+        expect(response.items[0].properties);
+        const title =
+          response.items[0].properties![view.space][
+            `${view.externalId}/${view.version}`
+          ]['title'].toString();
+        expect(title.startsWith('titl'));
       },
-      limit: 1,
-    });
-    expect(response.items).toHaveLength(1);
-    expect(response.items[0].properties);
-    const title =
-      response.items[0].properties![view.space][
-        `${view.externalId}/${view.version}`
-      ]['title'].toString();
-    expect(title.startsWith('titl'));
-  });
+      { interval: 1000, timeout: 30_000 }
+    );
+  }, 30_000);
 
   test('aggregate', async () => {
     const response = await client.instances.aggregate({