+++ packages/stable/src/__tests__/api/viewApi.int.spec.ts
@@ -1,9 +1,8 @@
 // Copyright 2020 Cognite AS
 
-import { ExternalView, View } from '../../types';
-import CogniteClient from '../../cogniteClient';
+import { afterAll, beforeAll, describe, expect, it } from 'vitest';
 import { setupLoggedInClient } from '../testUtils';
-import { randomInt } from '@cognite/sdk-core/src/testUtils';
+import { randomInt } from '../testUtils';
 
 describe('template view test', () => {
   let client: CogniteClient;