+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.10.2...@cognite/sdk-core@4.10.3) (2024-08-20)
+
+**Note:** Version bump only for package @cognite/sdk-core
+
+## [4.10.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.10.1...@cognite/sdk-core@4.10.2) (2024-08-19)
+
+**Note:** Version bump only for package @cognite/sdk-core
+
 ## [4.10.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.10.0...@cognite/sdk-core@4.10.1) (2024-01-22)
 
 ### Bug Fixes