+++ packages/stable/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.15.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.15.2...@cognite/sdk@9.15.3) (2024-08-20)
+
+### Bug Fixes
+
+- add missing properties to datapoints api ([#1124](https://github.com/cognitedata/cognite-sdk-js/issues/1124)) ([42c3160](https://github.com/cognitedata/cognite-sdk-js/commit/42c31609a7809ba866eeb62251fd765ab93eb7bc))
+
+## [9.15.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.15.1...@cognite/sdk@9.15.2) (2024-08-19)
+
+**Note:** Version bump only for package @cognite/sdk
+
 ## [9.15.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.15.0...@cognite/sdk@9.15.1) (2024-06-10)
 
 ### Bug Fixes